break out app group header component for readability
